server/requirements.txt
google-cloud-texttospeech
goose3<4.0.0,>=3.1.6
flask
flask_restful
mongoengine
flask-mongoengine
flask-marshmallow
requests
python-dateutil
redis
# Data and workers.
redis==3.5.3
celery==4.4.6
# Application server for both development and production.
gunicorn==20.0.4
